+++ src/redux/modules/language/selectors.js
@@ -1,0 +1,1 @@
+export const selectCurrentLocale = state => state.getIn(['Language', 'currentLocale']);