+++ CHANGELOG.md
@@ -28,6 +28,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Dispositions
 - Disposition Lists
 
+## [0.82.0] - 2020-08-20
+### Added
+- CXV1-22912 - Config 2 - Support Localization
+- CXV1-22914 - Config 2 - Add Translation to Navbar
+
 ## [0.81.29] - 2020-09-01
 ### Fixed
 - <no-Jira> - Update chinese & finnish language localization keys of contact attributes entity.